git-svn-id: svn+ssh://software-srn.sandia.gov/svn/spparks/trunk@666 685d9ffe-9724-0410-857b-ce06367cce3e

+++ src/app_lattice.h
@@ -14,6 +14,7 @@
 #ifndef SPK_APP_LATTICE_H
 #define SPK_APP_LATTICE_H
 
+#include "stdint.h"
 #include "stdio.h"
 #include "app.h"
 
@@ -71,7 +72,7 @@ class AppLattice : public App {
  protected:
   int me,nprocs;
 
-  int naccept,nattempt;       // number of accepted and attempted events
+  uint64_t naccept,nattempt;  // number of accepted and attempted events
   int nsweeps;                // number of sweeps performed
   double temperature,t_inverse;  // temperature settings
   double dt_sweep;            // rKMC time for nglobal attemped events
