+++ src/app_off_lattice.h
@@ -14,6 +14,7 @@
 #ifndef SPK_APP_OFF_LATTICE_H
 #define SPK_APP_OFF_LATTICE_H
 
+#include "stdint.h"
 #include "stdio.h"
 #include "app.h"
 
@@ -62,7 +63,7 @@ class AppOffLattice : public App {
   double subxlo,subylo,subzlo;
   double subxhi,subyhi,subzhi;
 
-  int naccept,nattempt;       // number of accepted and attempted events
+  uint64_t naccept,nattempt;  // number of accepted and attempted events
   int nsweeps;                // number of sweeps performed
   double temperature,t_inverse;  // temperature settings
   double dt_sweep;            // rKMC time for nglobal attemped events
